feat(ui): useController for custom radio buttons and extract EmissionsForm
So it's reusable for the value per subcategory section
Also rename SubSectorDrawer file to be PascalCase
