+++ app/src/app/[lng]/data/[step]/DirectMeasureForm.tsx
@@ -3,7 +3,6 @@ import {
   Box,
   FormControl,
   FormErrorMessage,
-  FormHelperText,
   FormLabel,
   HStack,
   Heading,
